requirements.txt
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
asyncpg==0.29.0
asyncstdlib==3.10.9
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
bs4==0.0.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
curl-cffi==0.5.10
dnspython==2.4.2
duckduckgo-search==3.9.8
environs==9.5.0
exceptiongroup==1.2.0
fake-useragent==1.4.0
fastapi==0.104.1
Flask==3.0.0
Flask-Admin==1.6.1
frozenlist==1.4.0
g4f==0.1.9.0
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
Js2Py==0.74
kombu==5.3.4
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
magic-filter==1.0.12
MarkupSafe==2.1.3
marshmallow==3.20.1
mmh3==4.0.1
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
py-arkose-generator==0.0.0.2
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PyExecJS==1.5.1
pyjsparser==2.7.1
pymongo==4.6.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
sqladmin==0.16.0
SQLAlchemy==1.4.0
starlette==0.27.0
tenacity==8.1.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver==3.5.4
urllib3==2.1.0
uvicorn==0.24.0.post1
vine==5.1.0
waitress==2.1.2
watchdog==3.0.0
wcwidth==0.2.12
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.1
yarl==1.9.3
